refactor(*): domain driven backend
