+++ .env_example
@@ -2,9 +2,12 @@
 CLIENT_PORT=8000
 # SERVER
 SERVER_PORT=3000
+SERVER_URL=http://localhost
 # DB
 MARIADB_DATABASE=matcha
 MARIADB_ROOT_PASSWORD=CHANGE_ME
 
 MARIADB_USER=matcha
 MARIADB_PASSWORD=CHANGE_ME
+
+JWT_SECRET=CHANGE_ME
