+++ server/package.json
@@ -15,7 +15,7 @@
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
-    "jsonwebtoken": "^9.0.0",
+    "jsonwebtoken": "^9.0.2",
     "mariadb": "^3.2.2",
     "serve": "^14.2.1",
     "uuid": "^9.0.0"
