+++ server/src/auth/auth.router.js
@@ -1,0 +1,9 @@
+import express from 'express';
+
+import authController from './auth.controller.js';
+
+const router = express.Router();
+
+router.post("/signup", authController.signup );
+
+export default router;
