--- server/src/routes/authRouter.js
@@ -1,8 +1,0 @@
-import express from 'express';
-import userController from '../controllers/userController.js';
-
-const router = express.Router();
-
-router.post("/signup", userController.signup );
-
-export default router;
