+++ server/src/user/user.controller.js
@@ -1,27 +1,4 @@
-import User from '../models/user.js';
-
-export async function signup(req, res) {
-    const email = req.body.email;
-    const password = req.body.password;
-
-    if (!email || !password) {
-        return res.status(400).json({ message: "Missing email or password" });
-    }
-
-    const user = new User(email, password);
-    const userExists = await user.getFromEmail();
-    
-    if (userExists) {
-        return res.status(400).json({ message: "Email already taken" });
-    }
-    try {
-        await user.saveToDB();
-        const newUser = await user.getFromEmail(email);
-        return res.status(201).json({ message: "User created", user_id: newUser.id });
-    } catch (error) {
-        return res.status(500).json({ message: error.message });
-    }
-}
+import User from './user.model.js';
 
 async function getAll(req, res) {
     const user = new User();
@@ -53,4 +30,4 @@ async function update(req, res) {
     }
 }
 
-export default { signup, getAll, getById, update };
+export default { getAll, getById, update };
