+++ server/src/user/user.router.js
@@ -1,5 +1,5 @@
 import express from 'express';
-import userController from '../controllers/userController.js';
+import userController from './user.controller.js';
 
 const router = express.Router();
